Do not set cookies on pages marked as cacheable
This applies to all cookies currently, with particular attention placed on the MV_SESSION_ID.  This
change is intended to alleviate the "first page hit" cache issue which leads to IC assigning a session
id even for pages which would be fully cacheable and then never actually generating the cacheable
content in question until a subsequent hit from a UA with the provided MV_SESSION_ID.

This is of particular concern when considering the case of a DDoS in which a huge number of connections
are made from unique instances.  Since the intent of the nginx cache is to prevent such traffic from
ever reaching the backend, the existing behavior would result in a counter-intuitive effect of all bot
traffic being passed directly to the "protected" backend when there was no document already in the nginx
cache, with the result that each backend-handled request would be itself uncacheable, having a Set-Cookie
header returned with the response.

Currently this fix will affect all cookies being generated from interchange on a page which is marked
cacheable (specifically one which has a Pragma cache_control value set to something other than no-cache).
We consider this acceptable as we are also moving the setting of the existing cookies to client-side code
where possible; if this ends up being untenable, we will see about changing this specific code to just
affect the MV_SESSION_ID cookie.